code cleanup

+++ StolenNetwork/PacketWriter.cs
@@ -75,7 +75,7 @@ namespace StolenNetwork
 				_peer.PacketStart();
 				_peer.PacketWrite(_stream);
 
-				return _peer.PacketBroadcast(info.Priority, info.Reliability, (byte) info.Channel);
+				return _peer.PacketBroadcast(info.Priority, info.Reliability, info.Channel);
 			}
 
 			if (info.Connections != null)
@@ -87,7 +87,7 @@ namespace StolenNetwork
 					_peer.PacketStart();
 					_peer.PacketWrite(_stream);
 
-					_peer.PacketSend(connection.Guid, info.Priority, info.Reliability, (byte) info.Channel);
+					_peer.PacketSend(connection.Guid, info.Priority, info.Reliability, info.Channel);
 				}
 
 				return 0;
@@ -98,7 +98,7 @@ namespace StolenNetwork
 				_peer.PacketStart();
 				_peer.PacketWrite(_stream);
 
-				return _peer.PacketSend(info.Connection.Guid, info.Priority, info.Reliability, (byte) info.Channel);
+				return _peer.PacketSend(info.Connection.Guid, info.Priority, info.Reliability, info.Channel);
 			}
 
 			return 0;
